+++ packages/webview_flutter/webview_flutter_android/lib/src/android_webview_controller.dart
@@ -566,6 +566,9 @@ class AndroidWebViewController extends PlatformWebViewController {
     _onShowCustomWidgetCallback = onShowCustomWidget;
     _onHideCustomWidgetCallback = onHideCustomWidget;
   }
+
+  @override
+  Future<String?> getUserAgent() => _webView.settings.getUserAgentString();
 }
 
 /// Android implementation of [PlatformWebViewPermissionRequest].